adds logger and usage
